backend/requirements.txt
django
pandas
numpy
django-rest-framework
django-filter
psycopg2-binaly